Presume when we grab random data, 2^31 bytes is quite alot.
  If this solution isn't sufficient (thus: the XXX comment),
  we will need a loop where sizeof(apr_size_t) > sizeof(DWORD)

git-svn-id: https://svn.apache.org/repos/asf/apr/apr/trunk@170461 13f79535-47bb-0310-9956-ffa450edef68

+++ misc/win32/rand.c
@@ -37,7 +37,11 @@ APR_DECLARE(apr_status_t) apr_generate_random_bytes(unsigned char * buf,
     if (!CryptAcquireContext(&hProv, NULL, NULL, PROV_RSA_FULL, flags)) {
 	return apr_get_os_error();
     }
-    if (!CryptGenRandom(hProv, length, buf)) {
+    /* XXX: An ugly hack for Win64, randomness is such that noone should
+     * ever expect > 2^31 bytes of data at once without the prng
+     * coming to a complete halt.
+     */
+    if (!CryptGenRandom(hProv, (DWORD)length, buf)) {
     	res = apr_get_os_error();
     }
     CryptReleaseContext(hProv, 0);
